TimePlayed virker nu, men den klager over at den ikke kan finde logging implementation

+++ src/GUI/Controller/mainMyTunesViewController.java
@@ -56,7 +56,7 @@ public class mainMyTunesViewController implements Initializable {
     @FXML
     private TableColumn<Song, String> colCategory = new TableColumn<>();
     @FXML
-    private TableColumn<Song, Integer> colTimeSongs = new TableColumn<>();
+    private TableColumn<Song, Double> colTimeSongs = new TableColumn<>();
     @FXML
     private TableColumn<Playlist, String> colName = new TableColumn<>();
     @FXML
